refactoring task handler to be more readable, removed an initialize dependency
