+++ app/graphql/tasker/queries/one_task.rb
@@ -12,7 +12,7 @@ module Tasker
 
       sig { params(task_id: T.any(Integer, String)).returns(Tasker::Task) }
       def resolve(task_id:)
-        task_query_base.where(task_id: task_id).first
+        Tasker::Task.with_all_associated.where(task_id: task_id).first
       end
     end
   end
