+++ app/graphql/tasker/queries/tasks_by_annotation.rb
@@ -19,7 +19,8 @@ module Tasker
       sig { params(limit: T.nilable(Integer), offset: T.nilable(Integer), sort_by: T.nilable(T.any(String, Symbol)), sort_order: T.nilable(T.any(String, Symbol)), annotation_type: T.any(String, Symbol), annotation_key: T.any(String, Symbol), annotation_value: T.any(String, Symbol)).returns(ActiveRecord::Relation) }
       def resolve(limit:, offset:, sort_by:, sort_order:, annotation_type:, annotation_key:, annotation_value:)
         sorts = page_sort_params(Task, limit, offset, sort_by, sort_order)
-        task_query_base
+        Tasker::Task
+          .with_all_associated
           .by_annotation(annotation_type, annotation_key, annotation_value)
           .limit(sorts[:limit])
           .offset(sorts[:offset])
